Lorentz.Contracts.Expiring

Documentation

data Parameter cp

Either a WrappedParameter, which will fail if the contract has expired, or GetExpiration (will never fail).

Constructors

WrappedParameter !cp

GetExpiration (View_ Timestamp)

data Storage st

We store the wrapped Contract's storage and the expiration Timestamp

Constructors

Storage

Fields

wrappedStorage :: !st

expirationTime :: !Timestamp

unStorage :: (Storage st & s) :-> ((st, Timestamp) & s)

forcedCoerce_ from Storage

toStorage :: ((st, Timestamp) & s) :-> (Storage st & s)

forcedCoerce_ to Storage

assertNotExpired :: (Timestamp & s) :-> (Timestamp & s)

If the Timestamp is not before now, throws "expired".

Caveat: now does not return the current time, it returns the timestamp of the last baked block. This means that now can be up to time_between_blocks behind realtime.

Currently, time_between_blocks is 60 seconds on mainnet or 30 seconds on testnet.

expiringContract :: forall cp st. IsoValue cp => ContractCode cp st -> ContractCode (Parameter cp) (Storage st)

Adds non-changeable expiration to a contract:

Storage contains an expiration Timestamp that can be queried using GetExpiration.

Once the Timestamp is past, the contract is locked

Caveat: Up to error due to now, see assertNotExpired for more info