Lorentz.Contracts.Forwarder

Documentation

data Storage

We need the addresses of: - The sub-token contract, assumed to accept Parameter - The Tezos Wallet to process refunds (assuming this contract is authorized to call it) - The central wallet to transfer sub-tokens to

Constructors

Storage

Fields

subTokenContract :: ContractRef Parameter

tezosWallet :: ContractRef RefundParameters

centralWallet :: Address

type RefundParameters = ("amount" :! Mutez, "to" :! Address)

What's required to refund mutez from the Tezos Wallet

type Parameter = Natural

The number of sub-tokens to transfer

toTransferParams :: (Address & (Address & (Natural & s))) :-> (TransferParams & s)

coerce_ to TransferParams

runTransferParams :: (ContractRef Parameter & (Address & (Address & (Natural & s)))) :-> (Operation & s)

Run TransferParams with a ContractRef Parameter, from Address, to Address, and number of sub-tokens

runStorageTransferParams :: (Parameter & (Storage & s)) :-> (Operation & (Storage & s))

Run TransferParams on the given Parameter and Storage, where from is sender and to is centralWallet

toRefundParameters :: (Mutez & (Storage & s)) :-> (RefundParameters & (Mutez & (ContractRef RefundParameters & (Storage & s))))

Derive RefundParameters and transfer arguments from the number of Mutez to refund and Storage

processRefund :: (Mutez & (Storage & s)) :-> (Operation & (Storage & s))

Process a refund, given the refund amount in Mutez and Storage

forwarderContract :: (forall s. (Natural & s) :-> (Mutez & s)) -> ContractCode Parameter Storage

Given a method to calculate the number of Mutez to refund from the number of sub-tokens transferred, produce a forwarder contract.

Orphan instances

Show Parameter

Instance details

Methods

showsPrec :: Int -> Parameter -> ShowS

show :: Parameter -> String

showList :: [Parameter] -> ShowS