Lorentz.Contracts.Forwarder.Specialized

Documentation

type Parameter = Natural

The number of sub-tokens to forward

type Storage = ()

We have the addresses of: - The sub-token contract, assumed to accept Parameter - The central wallet to transfer sub-tokens to

toTransferParameter :: forall s. (Address & (Natural & s)) :-> (TransferParams & s)

Wrap a to Address and number of tokens to transfer in TransferParams, sending from self

runSpecializedTransfer :: Address -> Address -> (Natural & s) :-> (Operation & s)

Run a transfer to the given central wallet Address, given the token contract Address and the number of tokens to transfer

specializedForwarderContract :: Address -> Address -> ContractCode Parameter Storage

Forwarder contract: forwards the given number of sub-tokens from its own address to the central wallet.

analyzeSpecializedForwarder :: Address -> Address -> AnalyzerRes

analyzeLorentz specialized to the specializedForwarderContract

contractOverValue :: forall cp st. ContractCode cp st -> ContractCode (Value (ToT cp)) (Value (ToT st))

forcedCoerce_ to convert parameter and storage types to their Value equivalents

verifyForwarderContract :: Address -> Address -> SomeContract -> Either String ()

Verify that SomeContract is an instance of specializedForwarderContract, for some particular central wallet address and token address.