Lorentz.Contracts.Forwarder.Specialized.FlushAny

Documentation

data Parameter

The number of sub-tokens to forward and the typed contract address of the token to forward it on

Constructors

Parameter

Fields

amountToFlush :: !Natural

tokenContract :: !(ContractRef TransferParams)

unParameter :: (Parameter & s) :-> ((Natural, ContractRef TransferParams) & s)

Unwrap a Parameter

mkParameter :: Natural -> Address -> Parameter

Assumes the Address points to a Parameter

type Storage = ()

We have the addresses of: - The central wallet to transfer sub-tokens to

toTransferParameter :: forall s. (Address & (Natural & s)) :-> (TransferParams & s)

Wrap a to Address and number of tokens to transfer in TransferParams, sending from self

runSpecializedAnyTransfer :: Address -> (Natural & (ContractRef TransferParams & s)) :-> (Operation & s)

Run a transfer to the given central wallet Address, given the token contract Address and the number of tokens to transfer

specializedAnyForwarderContract :: Address -> ContractCode Parameter Storage

Forwarder contract: forwards the given number of sub-tokens from its own address to the central wallet.

analyzeSpecializedAnyForwarder :: Address -> AnalyzerRes

analyzeLorentz specialized to the specializedAnyForwarderContract

verifyForwarderContract :: Address -> SomeContract -> Either String ()

Verify that SomeContract is an instance of specializedAnyForwarderContract, for some particular central wallet address and token address.