Lorentz.Contracts.Forwarder.Specialized.FlushAny.ForwardAnyFA12

Documentation

type Storage = ()

We have the addresses of: - The central wallet to transfer sub-tokens to

data Parameter

Constructors

Default ()

Flush Parameter

runSpecializedAnyTezTransfer :: Address -> (Maybe Parameter & s) :-> ([Operation] & s)

Run a transfer to the given central wallet Address, given Parameter. If no Parameter is provided, skip forwarding contract tokens.

Always attempt to forward any Tez in balance.

specializedAnyFA12ForwarderContract :: Address -> ContractCode Parameter Storage

Forwarder contract: forwards the given number of sub-tokens from its own address to the central wallet.

It also forwards all held Tez to the central wallet.

analyzeSpecializedAnyTezForwarder :: Address -> AnalyzerRes

analyzeLorentz specialized to the specializedAnyFA12ForwarderContract

verifyForwarderContract :: Address -> SomeContract -> Either String ()

Verify that SomeContract is an instance of specializedAnyFA12ForwarderContract, for some particular central wallet address and token address.

Orphan instances

SingI t => ParameterHasEntryPoints (Value t)

Instance details

Associated Types

type ParameterEntryPointsDerivation (Value t) :: Type