Lorentz.Contracts.Forwarder.Specialized.FlushAny.Tez

Documentation

type Storage = ()

We have the addresses of: - The central wallet to transfer sub-tokens to

runSpecializedAnyTezTransfer :: Address -> (Natural & (ContractRef TransferParams & s)) :-> ([Operation] & s)

Run a transfer to the given central wallet Address, given the number of tokens to forward and the token contract reference.

Always attempt to forward any Tez in balance.

specializedAnyTezForwarderContract :: Address -> ContractCode Parameter Storage

Forwarder contract: forwards the given number of sub-tokens from its own address to the central wallet.

It also forwards all held Tez to the central wallet.

analyzeSpecializedAnyTezForwarder :: Address -> AnalyzerRes

analyzeLorentz specialized to the specializedAnyTezForwarderContract

verifyForwarderContract :: Address -> SomeContract -> Either String ()

Verify that SomeContract is an instance of specializedAnyTezForwarderContract, for some particular central wallet address and token address.